feat: bottom navigation
