+++ src/screens/booked-vehicles.screen.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { Text } from "@ui-kitten/components";
+
+const BookedVehiclesScreen = () => {
+  return <Text>Bookings</Text>;
+};
+
+export default BookedVehiclesScreen;
