+++ src/screens/index.ts
@@ -1,13 +1,17 @@
 import Home from "./home.screen";
+import Search from "./search.screen";
 import VehicleDetail from "./vehicle-detail.screen";
 import SearchResult from "./search-result.screen";
-import Booking from "./booking.screen";
+import CreateBooking from "./create-booking.screen";
 import Info from "./info.screen";
+import BookedVehicles from "./booked-vehicles.screen";
 
 export default {
   Home,
+  Search,
   VehicleDetail,
   SearchResult,
-  Booking,
+  CreateBooking,
   Info,
+  BookedVehicles,
 };
